Split FileSystem.Statfs for Darwin
